src/requirements.txt
uvicorn
fastapi
sentry-sdk
arq
foxglove-web
aiofiles
ujson
sqlalchemy
psycopg2
markupsafe
chevron
phonenumbers
ua_parser
misaka
libsass
tqdm
uvicorn
pydantic[email]
jinja2
